requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
redis==5.0.1
celery==5.3.4 